better placement of the module classes
