+++ service/device/management/packages/internal/src/main/java/org/eclipse/kapua/service/device/management/packages/internal/DeviceManagementPackagesModule.java
@@ -1,0 +1,25 @@
+/*******************************************************************************
+ * Copyright (c) 2021 Eurotech and/or its affiliates and others
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     Eurotech - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.kapua.service.device.management.packages.internal;
+
+import org.eclipse.kapua.commons.core.AbstractKapuaModule;
+import org.eclipse.kapua.service.device.management.packages.DevicePackageFactory;
+import org.eclipse.kapua.service.device.management.packages.DevicePackageManagementService;
+
+public class DeviceManagementPackagesModule extends AbstractKapuaModule {
+    @Override
+    protected void configureModule() {
+        bind(DevicePackageFactory.class).to(DevicePackageFactoryImpl.class);
+        bind(DevicePackageManagementService.class).to(DevicePackageManagementServiceImpl.class);
+    }
+}
